fix of testweb errors

+++ geom/monetdb5/geom.c
@@ -4549,7 +4549,7 @@ int mbrWRITE(mbr *c, stream *s, size_t cnt) {
 /* return length of resulting string. */
 size_t wkbaTOSTR(char **toStr, int* len, wkba *fromArray) {
 	int items = fromArray->itemsNum, i;
-	double itemsNumDigits = ceil(log10(items));
+	int itemsNumDigits = (int)ceil(log10(items));
 	size_t dataSize;//, skipBytes=0;
 	char** partialStrs;
 	char* nilStr = "nil";
@@ -4627,7 +4627,7 @@ fprintf(stderr, "wkbaFROMSTR\n");
 	*toArray = (wkba*)GDKmalloc(wkba_size(items));
 
 	for(i=0; i<items; i++) {
-		int parsedBytes = wkbFROMSTR(fromStr+skipBytes, len, &((*toArray)->data[i]), srid);
+		size_t parsedBytes = wkbFROMSTR(fromStr+skipBytes, len, &((*toArray)->data[i]), srid);
 		skipBytes+=parsedBytes;
 	}
